move bruh that back to music.py
