+++ src/bruh.py
@@ -13,7 +13,6 @@ import numpy as np
 from io import BytesIO
 
 import verbosify
-from mediawikiapi import MediaWikiAPI
 
 from PyDictionary import PyDictionary
 dictionary = PyDictionary()
@@ -373,31 +372,6 @@ class Bruh(commands.Cog):
         r = requests.get("https://source.unsplash.com/featured/?cucumber")
         await ctx.send(r.url)
     
-
-
-
-    @commands.command(name='that')
-    async def that_(self, ctx, *args):
-        mediawikiapi = MediaWikiAPI()
-
-        # get random articles and number
-        rand_articles, rand_num = mediawikiapi.random(2), random.randint(0, 16777215)
-        article_md = ['[{}]({})'.format(article, 'https://en.wikipedia.org/wiki/'+article.replace(' ', '_')) for article in rand_articles]
-
-        # create embed
-        if not args or args[0] not in ['verbose', 'verbosify']: # zero or wrong arguments
-            embed = discord.Embed(color=discord.Color(rand_num), description='That is so {1}, can we hit {0} {2}'.format(rand_num, *article_md))
-        else: # either verbose or verbosify
-            embed = discord.Embed(color=discord.Color(rand_num), description='**That is so {1}, can we hit {0} {2}**'.format(rand_num, *article_md))
-            article_descriptions = [mediawikiapi.summary(article, chars=150, auto_suggest=False) for article in rand_articles]
-
-            if args[0] == 'verbose':
-                [embed.add_field(name="** **", value=desc, inline=True) for desc in article_descriptions]
-            elif args[0] == 'verbosify':
-                [embed.add_field(name="** **", value=verbosify._verbosify(desc), inline=True) for desc in article_descriptions]
-            
-
-        await ctx.send(embed=embed)
 
 
 def setup(bot):
